# openapi.yaml
openapi: 3.0.3
info:
  title: OScanner-Eng Partner API
  version: 1.0.1
  description: |
    Read-only access to OScanner-Eng user identity data and speaking-assessment
    results assembled from persisted question and recording identity snapshots.
    Responses intentionally exclude recorded videos, video URLs, filenames, MIME
    metadata, extracted audio paths, sampled-frame artifacts, transcripts, raw
    audio metrics, improved-answer text, and internal model metadata.
    Guest records are excluded from partner user exports. Learner endpoints accept
    either a signed DingTalk session or a signed guest cookie issued by GET /api/me.
    All learner mutations require X-Expected-Owner equal to the effective user.openId;
    a missing or mismatched value returns 409 with code IDENTITY_CHANGED before upload
    processing. The same header is optional on reads, where mismatches are also rejected.
servers:
  - url: /
    description: Current OScanner-Eng server
...
security:
  - bearerAuth: []
paths:
...
  /api/admission:
    post:
      summary: Enter or resume the FIFO waiting room
      description: Requires current privacy consent. One outstanding session per guest or DingTalk owner. Heartbeat every 20 seconds; unsubmitted reservations expire after three minutes.
      security: [{sessionCookie: []}, {guestCookie: []}]
      responses:
        '200':
          description: Admission with waiting, admitted, or submitted state and current wait estimate
          content:
            application/json:
              schema:
                type: object
                properties:
                  admission: {$ref: '#/components/schemas/Admission'}
        '429': {$ref: '#/components/responses/QueueBusy'}
    get:
      summary: Read the current owner's admission without creating one
      security: [{sessionCookie: []}, {guestCookie: []}]
      responses:
        '200':
          description: admission object or null
    delete:
      summary: Release an unsubmitted admission or waiting-room position
      security: [{sessionCookie: []}, {guestCookie: []}]
      responses:
        '200':
          description: Released; accepted jobs continue processing
  /api/admission/heartbeat:
    post:
      summary: Renew an existing admission lease
      security: [{sessionCookie: []}, {guestCookie: []}]
      responses:
        '200':
          description: Current admission, or null when expired
  /api/admission/upload-grant:
    post:
      summary: Reserve one of four upload transfer slots
      description: Grants expire after 30 seconds unless the upload starts. Requires admission and privacy consent.
      security: [{sessionCookie: []}, {guestCookie: []}]
      responses:
        '200':
          description: Single-use grant and expiresAt in Unix milliseconds
        '409':
          description: Admission expired or transfer already in progress
        '429': {$ref: '#/components/responses/QueueBusy'}
  /api/save-answer:
    post:
      summary: Durably accept an owned recorded answer for background evaluation
      description: The question must already belong to the owner. Inspection and MP4 normalization happen after HTTP 202. Retry the same submission ID to recover its existing job without duplicate scoring.
      security: [{sessionCookie: []}, {guestCookie: []}]
      parameters:
        - {$ref: '#/components/parameters/SubmissionId'}
        - {$ref: '#/components/parameters/UploadGrant'}
        - name: X-Question-Id
          in: header
          required: true
          schema: {type: string}
      requestBody: {$ref: '#/components/requestBodies/QueuedVideo'}
      responses: &QueuedUploadResponses
        '202':
          description: Original upload and job persisted; poll statusUrl for evaluation
          content:
            application/json:
              schema: {$ref: '#/components/schemas/QueuedJob'}
        '200':
          description: Existing terminal job returned for an idempotent retry
        '400':
          description: Unsupported upload, oversized media, or invalid question/submission identifier
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403':
          description: Current privacy consent required
        '409':
          description: Grant expired, ID unavailable, or submission discarded
        '429': {$ref: '#/components/responses/QueueBusy'}
        '503': {$ref: '#/components/responses/QueueBusy'}
  /api/evaluate-video:
    post:
      summary: Durably accept a standalone speech upload
      description: Uses the same admission, grant, and asynchronous status flow as recorded answers. Guest uploads remain private and never enter the public gallery.
      security: [{sessionCookie: []}, {guestCookie: []}]
      parameters:
        - {$ref: '#/components/parameters/SubmissionId'}
        - {$ref: '#/components/parameters/UploadGrant'}
      requestBody: {$ref: '#/components/requestBodies/QueuedVideo'}
      responses: *QueuedUploadResponses
  /api/jobs/{id}:
    get:
      summary: Read owner-only job progress or the final evaluation
      security: [{sessionCookie: []}, {guestCookie: []}]
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Poll every five seconds with jitter, or every twenty seconds in background tabs
          content:
            application/json:
              schema: {$ref: '#/components/schemas/QueuedJob'}
        '404':
          description: Unknown job or another owner's job
  /api/save-answer/{id}/cancel:
    post:
      summary: Discard an owned submission and prevent late results
      security: [{sessionCookie: []}, {guestCookie: []}]
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Cancellation persisted; active media processing stops before cleanup
  /api/admin/queue:
    get:
      summary: Read aggregate capacity, processing times, and recent usage
      security: [{sessionCookie: [], adminToken: []}]
      responses:
        '200':
          description: Aggregate queue metrics without recording content or identity fields
    post:
      summary: Pause or resume new admissions while accepted jobs continue
      security: [{sessionCookie: [], adminToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paused]
              properties:
                paused: {type: boolean}
      responses:
        '200':
          description: Updated queue metrics
  /api/admin/monitor:
    get:
      summary: Read resource alerts, monitor heartbeat, and recent DING delivery results
      security: [{sessionCookie: [], adminToken: []}]
      responses:
        '200':
          description: Monitor availability, maintenance expiry, resource sample, active alerts and latest 30 delivery states; excludes credentials and learner data
        '401':
          description: Authentication required
        '403':
          description: Administrator access required
...
components:
  securitySchemes:
    guestCookie:
      type: apiKey
      in: cookie
      name: englisheval_guest
    sessionCookie:
      type: apiKey
      in: cookie
      name: englisheval_session
    adminToken:
      type: apiKey
      in: header
      name: x-admin-access-token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: The server-side `PARTNER_API_KEY` value.
...
  parameters:
    SubmissionId:
      name: X-Submission-Id
      in: header
      required: true
      schema: {type: string, format: uuid}
    UploadGrant:
      name: X-Upload-Grant
      in: header
      required: true
      schema: {type: string, format: uuid}
...
  requestBodies:
    QueuedVideo:
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            required: [video]
            properties:
              video: {type: string, format: binary, description: Supported video up to 250 MiB}
              questionId: {type: string}
              submissionId: {type: string, format: uuid}
              startedAt: {type: string}
  responses:
    QueueBusy:
      description: Capacity temporarily unavailable; retain the recording and retry later
      headers:
        Retry-After:
          schema: {type: integer, example: 5}
      content:
        application/json:
          schema:
            type: object
            properties:
              error: {type: string}
              code: {type: string}
    Unauthorized:
      description: Missing or invalid bearer API key
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    WaitEstimate:
      type: object
      nullable: true
      description: Measured remaining-time range in seconds; null during cold start or unavailable processing
      properties:
        low: {type: integer}
        high: {type: integer}
    Admission:
      type: object
      properties:
        id: {type: string, format: uuid}
        state: {type: string, enum: [waiting, admitted, submitted]}
        jobId: {type: string, nullable: true}
        queuePosition: {type: integer}
        elapsedSeconds: {type: integer}
        estimatedRemainingSeconds: {$ref: '#/components/schemas/WaitEstimate'}
        delayed: {type: string, nullable: true}
        pollAfterSeconds: {type: integer}
    QueuedJob:
      type: object
      properties:
        id: {type: string, format: uuid}
        jobId: {type: string, format: uuid}
        state: {type: string, enum: [queued, processing, completed, failed, canceled]}
        stage: {type: string, enum: [normalized, media, transcription, scoring]}
        statusUrl: {type: string}
        queuePosition: {type: integer}
        elapsedSeconds: {type: integer}
        estimatedRemainingSeconds: {$ref: '#/components/schemas/WaitEstimate'}
        delayed: {type: string, nullable: true}
        pollAfterSeconds: {type: integer}
        path: {type: string, nullable: true}
        filename: {type: string, nullable: true}
        evaluation:
          type: object
          description: Existing normalized evaluation object, available on completion or failure
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
...